Add files via upload
Simple module for getting data from the SLC Airport's websites on all the arrivals for a day, group all arrivals into 15 minute intervals and produce a csv showing the times of the day with the most folks arriving at the airport. I wrote this when I was an Uber driver for a while, so I could make a more educated guess on when the best times to wait around the airport for a ride might be.